refactor: isolate desktop extension event handling
Move plugin and skill catalog mutation reducers out of the control socket helper and keep waiter resolution in focused extension event helpers.

Split plugin control-socket coverage into catalog and install event test files so the general event test no longer exceeds 1000 lines.
